+++ front-react/src/inquire/InquireAccont.js
@@ -1,7 +1,14 @@
 import React from "react";
 
 const InquireAccont = () => {
-  return <div></div>;
+  return (
+    <div style={{ backgroundColor: "#ffffff", minHeight: "100vh", padding: "2rem" }}>
+      <img
+        src="./Images/inquire_transfer/Inquire.png" 
+        alt=""
+        style={{ width: "100%", border: "1px solid #ddd", borderRadius: "8px" }}
+      />
+    </div>
+  );
 };
-
 export default InquireAccont;
