+++ front-react/src/inquire/InquireAssets.js
@@ -1,7 +1,14 @@
 import React from "react";
 
 const InquireAssets = () => {
-  return <div></div>;
+  return (
+    <div style={{ backgroundColor: "#ffffff", minHeight: "100vh", padding: "2rem" }}>
+      <img
+        src="./Images/inquire_transfer/InquireAssets.png" 
+        alt=""
+        style={{ width: "100%", border: "1px solid #ddd", borderRadius: "8px" }}
+      />
+    </div>
+  );
 };
-
 export default InquireAssets;
