+++ front-react/src/transfer/admin/MultiAdmin.js
@@ -1,7 +1,15 @@
 import React from "react";
 
 const MultiAdmin = () => {
-  return <div></div>;
+  return (
+    <div style={{ backgroundColor: "#ffffff", minHeight: "100vh", padding: "2rem" }}>
+      <img
+        src="./Images/inquire_transfer/LimitAdmin.png" 
+        alt=""
+        style={{ width: "100%", border: "1px solid #ddd", borderRadius: "8px" }}
+      />
+    </div>
+  );
 };
 
 export default MultiAdmin;
