+++ front-react/src/transfer/customer/TransLimit.js
@@ -1,7 +1,18 @@
 import React from "react";
 
 const TransLimit = () => {
-  return <div></div>;
+  return (
+    <div style={{ backgroundColor: "#ffffff", minHeight: "100vh", padding: "2rem" }}>
+      <div style={{ display: "flex", justifyContent: "center" }}>
+        <img
+          src="./Images/inquire_transfer/TransLimit.png"
+          alt="자동이체 화면"
+          style={{ width: "50%", border: "1px solid #ddd", borderRadius: "8px" }}
+        />
+      </div>
+    </div>
+  );
 };
+  
 
 export default TransLimit;
