+++ front-react/src/transfer/customer/TransMulti.js
@@ -1,7 +1,16 @@
 import React from "react";
 
 const TransMulti = () => {
-  return <div></div>;
+  return (
+    <div style={{ backgroundColor: "#ffffff", minHeight: "100vh", padding: "2rem" }}>
+      <div style={{ display: "flex", justifyContent: "center" }}>
+        <img
+          src="./Images/inquire_transfer/TransMulti.png"
+          alt="자동이체 화면"
+          style={{ width: "50%", border: "1px solid #ddd", borderRadius: "8px" }}
+        />
+      </div>
+    </div>
+  );
 };
-
 export default TransMulti;
