+++ front-react/src/transfer/customer/Transfer.js
@@ -1,7 +1,16 @@
 import React from "react";
 
 const Transfer = () => {
-  return <div></div>;
+  return (
+    <div style={{ backgroundColor: "#ffffff", minHeight: "100vh", padding: "2rem" }}>
+      <img
+        src="./Images/inquire_transfer/TransInstant.png" 
+        alt=""
+        style={{ width: "100%", border: "1px solid #ddd", borderRadius: "8px" }}
+      />
+    </div>
+  );
 };
+
 
 export default Transfer;
