+++ src/main/java/com/boot/sound/customer/CustomerRepository.java
@@ -1,6 +1,5 @@
-package com.boot.sound.customer.repository;
+package com.boot.sound.customer;
 
-import com.boot.sound.customer.dto.CustomerDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<CustomerDTO, String> {
